feat: add SASL security scheme types for use by Kafka specs
As discussed in https://github.com/asyncapi/bindings/pull/56 this
adds additional security scheme types. The motivation for adding
them is to enable description of secured Kafka clusters, however
the security protocols and mechanisms being added are not unique
to Kafka, so this commit adds them as generic security schemes so
they could be used by other protocols as well.

Contributes to: #466

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>