Update asyncapi.md
I may be interpreting this incorrectly, but when I see SUBSCRIBE, the application is going to subscribe to the channel and consume messages.  When I see PUBLISH, the application is going to produce a message and send it to a channel.  I think the definitions were switched in the docs.  This PR is to swap the definitions.